requirements.txt
streamlit
requests
matplotlib
plotly
python-dotenv
